refactor(connectors): extract shared IMAP bodystructure + mail-core helpers
The imap and protonmail connectors carried near-identical copies of the
BODYSTRUCTURE walkers (~86% duplicated) and the mail-core preview/limit/
address helpers (~35%). Hoist both into mcp-connectors/shared:

- shared/imap-bodystructure.ts — `extractAttachments` + `findTextPlainPart`
  returning a canonical `AttachmentMeta` (structurally identical to the
  per-connector attachment types). One shared test replaces the two
  duplicated connector test files.
- shared/imap-mail-core.ts — `clampLimit`, `capPreview`, `formatAddress`,
  `PREVIEW_*` constants, and the `MailAddress` shape. Both cores re-export
  these (keeping their connector-specific type names) so importers and
  existing core tests are unchanged.

Pure, dependency-free helpers; behavior identical. imap + protonmail
packages typecheck clean and all their tests pass.
